refactor:  reorganize perf analyzer
