+++ src/analyzers/patchers/__init__.py
@@ -1,0 +1,1 @@
+ATTACH_DIR = "src/analyzers/patchers/attachments/"
