+++ src/analyzers/patchers/attachments/empty.c
@@ -1,0 +1,1 @@
+void test_fun() {}
